+++ packages/react-components/react-avatar/src/components/AvatarGroup/renderAvatarGroup.tsx
@@ -1,6 +1,4 @@
 import * as React from 'react';
-import { Popover, PopoverTrigger } from '@fluentui/react-popover';
-import { Tooltip } from '@fluentui/react-tooltip';
 import { getSlots } from '@fluentui/react-utilities';
 import { AvatarGroupContext } from '../../contexts/AvatarGroupContext';
 import type { AvatarGroupState, AvatarGroupSlots } from './AvatarGroup.types';
@@ -10,27 +8,11 @@ import type { AvatarGroupState, AvatarGroupSlots } from './AvatarGroup.types';
  */
 export const renderAvatarGroup_unstable = (state: AvatarGroupState) => {
   const { slots, slotProps } = getSlots<AvatarGroupSlots>(state);
-  const { layout, size, nonOverflowAvatarsCount } = state;
+  const { layout, size } = state;
 
   return (
-    <AvatarGroupContext.Provider value={{ layout, size, nonOverflowAvatarsCount }}>
-      <slots.root {...slotProps.root}>
-        {state.root.children}
-        {state.hasOverflow && slots.overflowButton && slots.overflowContent && slots.overflowSurface && (
-          <Popover trapFocus size="small">
-            <PopoverTrigger>
-              <Tooltip content={state.tooltipContent} relationship="label">
-                <slots.overflowButton {...slotProps.overflowButton} />
-              </Tooltip>
-            </PopoverTrigger>
-            <slots.overflowSurface {...slotProps.overflowSurface}>
-              <AvatarGroupContext.Provider value={{ isOverflow: true, layout, size: 24 }}>
-                <slots.overflowContent {...slotProps.overflowContent} />
-              </AvatarGroupContext.Provider>
-            </slots.overflowSurface>
-          </Popover>
-        )}
-      </slots.root>
+    <AvatarGroupContext.Provider value={{ layout, size }}>
+      <slots.root {...slotProps.root} />
     </AvatarGroupContext.Provider>
   );
 };
