+++ packages/react-components/react-avatar/src/components/AvatarGroupItem/AvatarGroupItem.types.ts
@@ -33,7 +33,6 @@ export type AvatarGroupItemState = ComponentState<AvatarGroupItemSlots> & {
    */
   isOverflowItem?: boolean;
 
-  nonOverflowAvatarsCount: number;
   layout: AvatarGroupProps['layout'];
   size: AvatarSizes;
 };
