+++ packages/react-components/react-avatar/src/components/AvatarGroupItem/useAvatarGroupItemStyles.ts
@@ -1,9 +1,9 @@
-import { useSizeStyles } from '../../Avatar';
 import { makeStyles, mergeClasses, shorthands } from '@griffel/react';
+import { useSizeStyles } from '../../Avatar';
 import { tokens, typographyStyles } from '@fluentui/react-theme';
-import type { AvatarSizes } from '../../Avatar';
-import type { AvatarGroupProps } from '../../AvatarGroup';
 import type { AvatarGroupItemSlots, AvatarGroupItemState } from './AvatarGroupItem.types';
+import type { AvatarGroupProps } from '../../AvatarGroup';
+import type { AvatarSizes } from '../../Avatar';
 import type { SlotClassNames } from '@fluentui/react-utilities';
 
 export const avatarGroupItemClassNames: SlotClassNames<AvatarGroupItemSlots> = {
@@ -11,8 +11,6 @@ export const avatarGroupItemClassNames: SlotClassNames<AvatarGroupItemSlots> = {
   avatar: 'fui-AvatarGroupItem__avatar',
   overflowLabel: 'fui-AvatarGroupItem__overflowLabel',
 };
-
-const avatarGroupItemDividerWidthVar = '--fuiAvatarGroupItem__divier--width';
 
 /**
  * Styles for the root slot
@@ -39,9 +37,6 @@ const useAvatarStyles = makeStyles({
   nonOverflowItem: {
     position: 'absolute',
   },
-  pie: {
-    ...shorthands.borderRadius(0),
-  },
 });
 
 /**
@@ -53,45 +48,6 @@ const useOverflowLabelStyles = makeStyles({
     color: tokens.colorNeutralForeground1,
     ...typographyStyles.body1,
   },
-});
-
-/**
- * Styles for the pie layout
- */
-const usePieStyles = makeStyles({
-  base: {
-    position: 'absolute',
-  },
-  twoSlices: {
-    '&:first-child': {
-      clipPath: `inset(0 calc(25% + (var(${avatarGroupItemDividerWidthVar}) / 2)) 0 25%)`,
-      left: '-25%',
-    },
-    '&:nth-child(2)': {
-      clipPath: `inset(0 25% 0 calc(25% + (var(${avatarGroupItemDividerWidthVar}) / 2)))`,
-      left: '25%',
-    },
-  },
-  threeSlices: {
-    '&:first-child': {
-      clipPath: `inset(0 calc(25% + (var(${avatarGroupItemDividerWidthVar}) / 2)) 0 25%)`,
-      left: '-25%',
-    },
-    '&:not(:first-child)': {
-      // Since the two AvatarGroupItems on the right are scaled by 0.5, the divider width should not be halved.
-      clipPath: `inset(0 0 var(${avatarGroupItemDividerWidthVar}) var(${avatarGroupItemDividerWidthVar}))`,
-      left: '50%',
-      transform: 'scale(0.5)',
-      transformOrigin: '0 0',
-    },
-    '&:nth-child(3)': {
-      clipPath: `inset(var(${avatarGroupItemDividerWidthVar}) 0 0 var(${avatarGroupItemDividerWidthVar}))`,
-      top: '50%',
-    },
-  },
-  thick: { [avatarGroupItemDividerWidthVar]: tokens.strokeWidthThick },
-  thicker: { [avatarGroupItemDividerWidthVar]: tokens.strokeWidthThicker },
-  thickest: { [avatarGroupItemDividerWidthVar]: tokens.strokeWidthThickest },
 });
 
 const useStackStyles = makeStyles({
@@ -111,15 +67,8 @@ const useStackStyles = makeStyles({
     },
   },
   overflowButton: {
-    // border-color has to be set to transparent when there's focus due to the outline overlapping the focus ring.
-    '&:focus': {
-      '&::after': {
-        ...shorthands.borderColor('transparent'),
-      },
-    },
-    // hide inner border when using high contrast mode and use the outer (::after) to match Avatar's outline
-    '@media (forced-colors: active)': {
-      ...shorthands.borderColor('Canvas'),
+    '&:focus::after': {
+      ...shorthands.borderColor('transparent'),
     },
   },
   thick: {
@@ -199,10 +148,9 @@ const useSpreadStyles = makeStyles({
  * Apply styling to the AvatarGroupItem slots based on the state
  */
 export const useAvatarGroupItemStyles_unstable = (state: AvatarGroupItemState): AvatarGroupItemState => {
-  const { nonOverflowAvatarsCount, isOverflowItem, layout, size } = state;
+  const { isOverflowItem, layout, size } = state;
 
   const rootStyles = useRootStyles();
-  const pieStyles = usePieStyles();
   const sizeStyles = useSizeStyles();
   const overflowLabelStyles = useOverflowLabelStyles();
   const avatarStyles = useAvatarStyles();
@@ -215,24 +163,6 @@ export const useAvatarGroupItemStyles_unstable = (state: AvatarGroupItemState): 
     rootClasses.push(rootStyles.nonOverflowItem);
     rootClasses.push(groupChildClassName);
     rootClasses.push(sizeStyles[size]);
-
-    if (layout === 'pie') {
-      rootClasses.push(pieStyles.base);
-
-      if (size < 56) {
-        rootClasses.push(pieStyles.thick);
-      } else if (size < 72) {
-        rootClasses.push(pieStyles.thicker);
-      } else {
-        rootClasses.push(pieStyles.thickest);
-      }
-
-      if (nonOverflowAvatarsCount === 2) {
-        rootClasses.push(pieStyles.twoSlices);
-      } else if (nonOverflowAvatarsCount === 3) {
-        rootClasses.push(pieStyles.threeSlices);
-      }
-    }
   } else {
     rootClasses.push(rootStyles.overflowItem);
   }
@@ -242,7 +172,6 @@ export const useAvatarGroupItemStyles_unstable = (state: AvatarGroupItemState): 
   state.avatar.className = mergeClasses(
     avatarGroupItemClassNames.avatar,
     !isOverflowItem && avatarStyles.nonOverflowItem,
-    !isOverflowItem && layout === 'pie' && avatarStyles.pie,
     state.avatar.className,
   );
