+++ packages/react-components/react-avatar/src/components/AvatarGroupOverflow/AvatarGroupOverflow.test.tsx
@@ -1,5 +1,3 @@
-import * as React from 'react';
-import { render } from '@testing-library/react';
 import { AvatarGroupOverflow } from './AvatarGroupOverflow';
 import { isConformant } from '../../common/isConformant';
 
@@ -7,12 +5,5 @@ describe('AvatarGroupOverflow', () => {
   isConformant({
     Component: AvatarGroupOverflow,
     displayName: 'AvatarGroupOverflow',
-  });
-
-  // TODO add more tests here, and create visual regression tests in /apps/vr-tests
-
-  it('renders a default state', () => {
-    const result = render(<AvatarGroupOverflow>Default AvatarGroupOverflow</AvatarGroupOverflow>);
-    expect(result.container).toMatchSnapshot();
   });
 });
