+++ packages/react-components/react-avatar/src/components/AvatarGroupOverflow/renderAvatarGroupOverflow.tsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
+import { AvatarGroupContext } from '../../contexts/AvatarGroupContext';
 import { getSlots } from '@fluentui/react-utilities';
+import { Popover, PopoverTrigger } from '@fluentui/react-popover';
+import { Tooltip } from '@fluentui/react-tooltip';
 import type { AvatarGroupOverflowState, AvatarGroupOverflowSlots } from './AvatarGroupOverflow.types';
 
 /**
@@ -8,6 +11,20 @@ import type { AvatarGroupOverflowState, AvatarGroupOverflowSlots } from './Avata
 export const renderAvatarGroupOverflow_unstable = (state: AvatarGroupOverflowState) => {
   const { slots, slotProps } = getSlots<AvatarGroupOverflowSlots>(state);
 
-  // TODO Add additional slots in the appropriate place
-  return <slots.root {...slotProps.root} />;
+  return (
+    <Popover trapFocus size="small">
+      <PopoverTrigger>
+        <Tooltip content={state.tooltipContent} relationship="label">
+          <slots.root {...slotProps.root} />
+        </Tooltip>
+      </PopoverTrigger>
+      {slots.overflowSurface && (
+        <slots.overflowSurface {...slotProps.overflowSurface}>
+          <AvatarGroupContext.Provider value={{ isOverflow: true, size: 24 }}>
+            {slots.overflowContent && <slots.overflowContent {...slotProps.overflowContent} />}
+          </AvatarGroupContext.Provider>
+        </slots.overflowSurface>
+      )}
+    </Popover>
+  );
 };
