+++ packages/react-components/react-avatar/src/contexts/AvatarGroupContext.types.ts
@@ -2,5 +2,4 @@ import type { AvatarGroupProps } from '../AvatarGroup';
 
 export type AvatarGroupContextValue = Pick<AvatarGroupProps, 'size' | 'layout'> & {
   isOverflow?: boolean;
-  nonOverflowAvatarsCount?: number;
 };
