+++ packages/react-components/react-avatar/src/contexts/index.ts
@@ -1,1 +1,2 @@
 export * from './AvatarGroupContext';
+export * from './AvatarGroupContext.types';
