+++ packages/react-components/react-avatar/src/stories/AvatarGroup/AvatarGroupDefault.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { AvatarGroup, AvatarGroupItem } from '@fluentui/react-avatar';
+import { AvatarGroup, AvatarGroupItem, AvatarGroupOverflow } from '@fluentui/react-avatar';
 import type { AvatarGroupProps } from '@fluentui/react-avatar';
 
 export const Default = (props: Partial<AvatarGroupProps>) => {
@@ -9,11 +9,13 @@ export const Default = (props: Partial<AvatarGroupProps>) => {
       <AvatarGroupItem name="Elvia Atkins" />
       <AvatarGroupItem name="Cameron Evans" />
       <AvatarGroupItem name="Wanda Howard" />
-      <AvatarGroupItem name="Mona Kane" />
-      <AvatarGroupItem name="Allan Munger" />
-      <AvatarGroupItem name="Daisy Phillips" />
-      <AvatarGroupItem name="Robert Tolbert" />
-      <AvatarGroupItem name="Kevin Sturgis" />
+      <AvatarGroupOverflow>
+        <AvatarGroupItem name="Mona Kane" />
+        <AvatarGroupItem name="Allan Munger" />
+        <AvatarGroupItem name="Daisy Phillips" />
+        <AvatarGroupItem name="Robert Tolbert" />
+        <AvatarGroupItem name="Kevin Sturgis" />
+      </AvatarGroupOverflow>
     </AvatarGroup>
   );
 };
