--- packages/react-components/react-avatar/src/stories/AvatarGroup/AvatarGroupMaxAvatars.stories.tsx
@@ -1,51 +1,0 @@
-import * as React from 'react';
-import { AvatarGroup, AvatarGroupItem } from '@fluentui/react-avatar';
-import { makeStyles } from '@griffel/react';
-
-const useStyles = makeStyles({
-  root: {
-    display: 'grid',
-    flexDirection: 'column',
-    rowGap: '10px',
-  },
-});
-
-export const MaxAvatars = () => {
-  const styles = useStyles();
-
-  return (
-    <div className={styles.root}>
-      <AvatarGroup maxAvatars={8}>
-        <AvatarGroupItem name="Katri Athokas" />
-        <AvatarGroupItem name="Elvia Atkins" />
-        <AvatarGroupItem name="Cameron Evans" />
-        <AvatarGroupItem name="Wanda Howard" />
-        <AvatarGroupItem name="Mona Kane" />
-        <AvatarGroupItem name="Allan Munger" />
-        <AvatarGroupItem name="Daisy Phillips" />
-        <AvatarGroupItem name="Robert Tolbert" />
-        <AvatarGroupItem name="Kevin Sturgis" />
-      </AvatarGroup>
-      <AvatarGroup maxAvatars={3}>
-        <AvatarGroupItem name="Katri Athokas" />
-        <AvatarGroupItem name="Elvia Atkins" />
-        <AvatarGroupItem name="Cameron Evans" />
-        <AvatarGroupItem name="Wanda Howard" />
-        <AvatarGroupItem name="Mona Kane" />
-        <AvatarGroupItem name="Allan Munger" />
-        <AvatarGroupItem name="Daisy Phillips" />
-        <AvatarGroupItem name="Robert Tolbert" />
-        <AvatarGroupItem name="Kevin Sturgis" />
-      </AvatarGroup>
-    </div>
-  );
-};
-
-MaxAvatars.parameters = {
-  docs: {
-    description: {
-      story: `An AvatarGroup supports a custom number of AvatarGroupItems to render before overflowing.
-      The default is 5.`,
-    },
-  },
-};
