+++ packages/react-components/react-avatar/src/stories/AvatarGroup/index.stories.tsx
@@ -7,8 +7,6 @@ export { Default } from './AvatarGroupDefault.stories';
 export { Layout } from './AvatarGroupLayout.stories';
 export { SizeSpread } from './AvatarGroupSizeSpread.stories';
 export { SizeStack } from './AvatarGroupSizeStack.stories';
-export { SizePie } from './AvatarGroupSizePie.stories';
-export { MaxAvatars } from './AvatarGroupMaxAvatars.stories';
 export { Indicator } from './AvatarGroupIndicator.stories';
 export { Color } from './AvatarGroupColor.stories';
