--- packages/react-components/react-avatar/src/stories/AvatarGroupOverflow/AvatarGroupOverflowBestPractices.md
@@ -1,5 +1,0 @@
-## Best practices
-
-### Do
-
-### Don't
