--- packages/react-components/react-avatar/src/stories/AvatarGroupOverflow/AvatarGroupOverflowDefault.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { AvatarGroupOverflow, AvatarGroupOverflowProps } from '@fluentui/react-avatar';
-
-export const Default = (props: Partial<AvatarGroupOverflowProps>) => <AvatarGroupOverflow {...props} />;
